pyproject.toml
[project]
name = "mister-agent-memory"
version = "1.0.0"
description = "Personal knowledge base compiled from AI conversations — with Obsidian sync, auto-tagging, and weekly digests"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.29",
    "python-dotenv>=1.0.0",
    "tzdata>=2024.1",
]

[tool.ruff]
line-length = 100
